fix: Move packages to internal/ and overhaul build system
Move all top-level packages (chromium, crocochrome, http, metrics,
testutil) under internal/, making them unexported to external
consumers. Rename cmd/main.go to cmd/crocochrome/main.go so the
binary name is derived from the directory as convention requires.
Update all import paths accordingly.

Overhaul the Makefile:

- Add LOCAL=true mode so targets can run natively (no Docker wrapper)
  when set explicitly or when CI=true is detected.
- Change build output path from build/sm-k6-archiver to
  dist/crocochrome.
- Add test-integration target for running integration tests directly.
- Add help target with per-target descriptions.
- Add clean/distclean targets.
- Remove stale Drone CI generation target.

Switch the Dockerfile build step from a raw go build invocation to
`make ... build`, so the container build stays in sync with the
Makefile definition.

Replace the root .gitignore (which only tracked build/) with a
dist/.gitignore that ignores dist/ contents while keeping the
directory itself trackable.

Add DEVELOPMENT.md documenting commands, architecture, package
layout, session lifecycle, and container setup.
